speech-to-text/requirements.txt
pyperclip==1.9.0
streamlit==1.26.0
yt-dlp
openai-whisper
pytorch
